Removed Boost

+++ Project/src/Common/Base/Types.h
@@ -4,7 +4,7 @@
 #include <string>
 
 //A few string things. StringID is unique identifier based on a string hash.
-typedef int StringID;
+typedef size_t StringID;
 //Easier use of string without calling std:: everywhere.
 typedef std::string String;
